Add login
Summary:
Add login page and session management.
When the user reaches the dashboard, if no session token is detected the user is redirected to the login page.
When the user successfuly logins, a session token is returned from the server, and this is persisted in both the sessionStore and the Session factory.
Session store is cleared when the browser tab is closed

Test Plan: N/A

Reviewers: tnu

Reviewed By: tnu

Differential Revision: http://phab.ailurus.ca/D45